+++ src/main/java/duke/DukeException.java
@@ -1,11 +1,13 @@
+package duke;
+
 /**
- * Custom exception class for Duke chatbot.
+ * Custom exception class for duke.Duke chatbot.
  * This exception is thrown when there is an issue or error encountered during chatbot operations.
  */
 public class DukeException extends Exception {
 
     /**
-     * Constructs a new DukeException with the specified detail message.
+     * Constructs a new duke.DukeException with the specified detail message.
      *
      * @param message The detail message explaining the reason for the exception.
      */
@@ -14,7 +16,7 @@ public class DukeException extends Exception {
     }
 
     /**
-     * Returns a string representation of this DukeException, which includes the error message.
+     * Returns a string representation of this duke.DukeException, which includes the error message.
      *
      * @return A string representation of the exception's detail message.
      */
