+++ src/main/java/duke/task/Event.java
@@ -1,5 +1,7 @@
+package duke.task;
+
 /**
- * The Event class represents a task of type "Event" inherited from the Task class.
+ * The duke.task.Event class represents a duke.task of type "duke.task.Event" inherited from the duke.task.Task class.
  * It contains a description and a time interval during which the event occurs.
  */
 public class Event extends Task {
@@ -7,9 +9,9 @@ public class Event extends Task {
     protected String end;
 
     /**
-     * Constructs a new Event task with the provided description and time interval.
+     * Constructs a new duke.task.Event duke.task with the provided description and time interval.
      *
-     * @param description The description of the Event task.
+     * @param description The description of the duke.task.Event duke.task.
      * @param start       The start time of the event.
      * @param end         The end time of the event.
      */
@@ -32,9 +34,9 @@ public class Event extends Task {
     }
 
     /**
-     * Returns a string representation of the Event task, including its completion status, description, and time interval.
+     * Returns a string representation of the duke.task.Event duke.task, including its completion status, description, and time interval.
      *
-     * @return A formatted string indicating the task type, completion status, and time interval.
+     * @return A formatted string indicating the duke.task type, completion status, and time interval.
      */
     @Override
     public String toString() {
